requirements.txt
smc-python>=0.5.6